api/requirements.fastapi.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=1.10.0
python-dotenv>=1.0.0
scikit-learn==1.5.2
pandas>=2.0.0
numpy>=1.24.0
mlflow==2.15.0
boto3>=1.28.0
xgboost>=1.7.0
joblib>=1.3.0